# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API Backend NeoPay - Sistem Antrian Digital Bank
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /daftar:
    post:
      operationId: daftar
      tags: [auth]
      summary: Daftar akun baru
      description: Mendaftarkan nasabah baru ke sistem
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DaftarBody"
      responses:
        "201":
          description: Berhasil daftar
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Data tidak valid
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /masuk:
    post:
      operationId: masuk
      tags: [auth]
      summary: Login
      description: Login dengan email dan password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MasukBody"
      responses:
        "200":
          description: Berhasil login
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Email atau password salah
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        pesan:
          type: string
        error:
          type: string
      required:
        - pesan
...
    DaftarBody:
      type: object
      properties:
        nama:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
        noHp:
          type: string
        nik:
          type: string
          nullable: true
        tanggalLahir:
          type: string
          description: Tanggal lahir format YYYY-MM-DD
      required:
        - nama
        - email
        - password
        - noHp
        - tanggalLahir
    MasukBody:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required:
        - email
        - password
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        nasabah:
          $ref: "#/components/schemas/NasabahResponse"
      required:
        - token
        - nasabah
    NasabahResponse:
      type: object
      properties:
        id:
          type: string
        supabaseUserId:
          type: string
        nama:
          type: string
        noHp:
          type: string
        email:
          type: string
        nik:
          type: string
          nullable: true
        noRekening:
          type: string
          nullable: true
        tanggalLahir:
          type: string
          nullable: true
        saldo:
          type: integer
          description: Saldo nasabah dalam rupiah
        role:
          type: string
          enum: [nasabah, teller, admin]
        oneSignalPlayerId:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - supabaseUserId
        - nama
        - noHp
        - email
        - saldo
        - role
        - createdAt
...