# lib/api-zod/src/generated/types/nasabahResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API Backend NeoPay - Sistem Antrian Digital Bank
 * OpenAPI spec version: 0.1.0
 */
import type { NasabahResponseRole } from "./nasabahResponseRole";

export interface NasabahResponse {
  id: string;
  supabaseUserId: string;
  nama: string;
  noHp: string;
  email: string;
  nik?: string | null;
  noRekening?: string | null;
  tanggalLahir?: string | null;
  /** Saldo nasabah dalam rupiah */
  saldo: number;
  role: NasabahResponseRole;
  oneSignalPlayerId?: string | null;
  createdAt: Date;
}